// src/api/model/filterOperation.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type FilterOperation = 'CONTAINS' | 'EQUALS' | 'GREATER_THAN' | 'LESS_THAN';

export const FilterOperation = {
    Contains: 'CONTAINS' as FilterOperation,
    Equals: 'EQUALS' as FilterOperation,
    GreaterThan: 'GREATER_THAN' as FilterOperation,
    LessThan: 'LESS_THAN' as FilterOperation
};